Refactored server.ts to separate the app-related logic with the app.listen function, to allow import app(without app.listen) in the test files.
