+++ server/src/db/investWithFriendsDb.ts
@@ -1,8 +1,6 @@
 import { Pool, QueryResult, QueryResultRow } from 'pg';
 import { PG_URI, PG_URI_TEST } from '../utils/config';
 
-console.log('PG_URI: ', PG_URI);
-
 const pool = new Pool({
   connectionString: process.env.NODE_ENV === 'test' ? PG_URI_TEST : PG_URI,
 });
@@ -15,7 +13,7 @@ export default {
     params?: Array<string | number>
     // callback: (err: Error, result: QueryResult<any>) => void
   ): Promise<QueryResult<QueryResultRow>> => {
-    console.log('Executed Query: ', text);
+    // console.log('Executed Query: ', text);
     return pool.query<QueryResultRow, Array<string | number>>(text, params);
   },
 };
